OmniSharpDocumentationCommentOptionsWrapper
