+++ src/EditorFeatures/VisualBasicTest/Wrapping/AbstractParameterWrappingTests.vb
@@ -6,6 +6,7 @@ Imports System.Collections.Immutable
 Imports Microsoft.CodeAnalysis.CodeActions
 Imports Microsoft.CodeAnalysis.Editor.UnitTests.CodeActions
 Imports Microsoft.CodeAnalysis.Editor.VisualBasic.UnitTests.CodeRefactorings
+Imports Microsoft.CodeAnalysis.ImplementType
 Imports Microsoft.CodeAnalysis.SymbolSearch
 
 Namespace Microsoft.CodeAnalysis.Editor.VisualBasic.UnitTests.Wrapping
@@ -17,7 +18,10 @@ Namespace Microsoft.CodeAnalysis.Editor.VisualBasic.UnitTests.Wrapping
         End Function
 
         Private Protected Shared Function GetIndentionColumn(column As Integer) As CodeActionOptions
-            Return New CodeActionOptions(SymbolSearchOptions.Default, WrappingColumn:=column)
+            Return New CodeActionOptions(
+                SymbolSearchOptions.Default,
+                ImplementTypeOptions.Default,
+                WrappingColumn:=column)
         End Function
 
         Protected Function TestAllWrappingCasesAsync(
